canister stabilization feat

+++ game_bloc_backend/src/model.rs
@@ -33,11 +33,11 @@ pub struct UserProfile {
 // The `Storable` trait is already implemented for several common types (e.g. u64),
 // so you can use those directly without implementing the `Storable` trait for them.
 impl Storable for UserProfile {
-    fn to_bytes(&self) -> std::borrow::Cow<[u8]> {
+    fn to_bytes(&self) -> std::borrow::Cow<[u128]> {
         Cow::Owned(Encode!(self).unwrap())
     }
 
-    fn from_bytes(bytes: std::borrow::Cow<[u8]>) -> Self {
+    fn from_bytes(bytes: std::borrow::Cow<[u128]>) -> Self {
         Decode!(bytes.as_ref(), Self).unwrap()
     }
